# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a list of posts",
                "description": "Returns a list of posts",
                "operationId": "7b253f6aade6833019f938c1f2ef08ed",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of posts per page (default is 5)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page (default is 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response: List of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "posts": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/PostResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden: Insufficient permissions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found: Posts not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Store a new post",
                "description": "Stores a new post in the database",
                "operationId": "b996500c6caf55100cd024d3e3b49dd5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "post": {
                                                    "$ref": "#/components/schemas/PostResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "description": "Validation error details",
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The title field is required."
                                                ],
                                                "slug": [
                                                    "The slug field is required.",
                                                    "The slug has already been taken."
                                                ],
                                                "body": [
                                                    "The body field is required."
                                                ],
                                                "author_id": [
                                                    "The user id field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePostRequest": {
                "title": "Store Post Request",
                "description": "Request structure for storing a new post",
                "required": [
                    "title",
                    "slug",
                    "body",
                    "author_id"
                ],
                "properties": {
                    "title": {
                        "description": "The title of the post",
                        "type": "string"
                    },
                    "slug": {
                        "description": "The slug of the post",
                        "type": "string"
                    },
                    "body": {
                        "description": "The body of the post",
                        "type": "string"
                    },
                    "author_id": {
                        "description": "The updated user ID associated with the post",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "AuthorResource": {
                "title": "Author Resource",
                "description": "Author Resource",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the author",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the author",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PostResource": {
                "title": "Post Resource",
                "description": "Post Resource",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the post",
                        "type": "integer"
                    },
                    "title": {
                        "description": "The title of the post",
                        "type": "string"
                    },
                    "slug": {
                        "description": "The slug of the post",
                        "type": "string"
                    },
                    "body": {
                        "description": "The body of the post",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The timestamp when the post was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "author": {
                        "$ref": "#/components/schemas/AuthorResource"
                    },
                    "updated_at": {
                        "description": "The timestamp when the post was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}